Fixed the !get command
The issue was caused by me changing the notation section to "description" and although it already regexped that it didn't the second time around. Also the YYG Wiki function needs changed to the online manual link since the Wiki is now down and replaced with the knowledge base, any of the old links would be a 404.